[SKYCOIN-31] [Mesh] Rename modules in mesh, to improve clarity
