+++ src/mesh/domain/message.go
@@ -86,11 +86,6 @@ type AddNodeMessage struct {
 //	MessageBase
 //}
 
-type RouteConfig struct {
-	Id    uuid.UUID
-	Peers []cipher.PubKey
-}
-
 type MessageToSend struct {
 	ThruRoute uuid.UUID
 	Contents  []byte
@@ -101,9 +96,4 @@ type MessageToReceive struct {
 	Contents      []byte
 	Reply         []byte
 	ReplyReliably bool
-}
-
-type ToConnect struct {
-	Peer cipher.PubKey
-	Info string
 }
