+++ src/mesh/domain/peer.go
@@ -1,0 +1,8 @@
+package domain
+
+import "github.com/skycoin/skycoin/src/cipher"
+
+type Peer struct {
+	Peer cipher.PubKey
+	Info string
+}
